pyproject.toml
[tool.poetry]
name = "zdppy_grpc"
version = "0.1.0"
description = ""
authors = ["zhangdapeng520 <pygosuperman@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.9"
grpcio = "^1.45.0"
grpcio-tools = "^1.45.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
